refactor: one storage type per file
Move out `FilterHeadersStorage` and `PersistentFilterHeadersStorage` into `storage/filters.rs` to have one storage type per file.
